Updated movie modal and added to card movie
